+++ ui/mira/src/pages/vulnerabilities.tsx
@@ -10,6 +10,7 @@ import {
   CardHeader,
   CardTitle,
 } from "@/components/ui/card"
+import { Skeleton } from "@/components/ui/skeleton"
 import {
   Table,
   TableBody,
@@ -154,7 +155,11 @@ export function VulnerabilitiesPage() {
           <div className="flex items-center justify-between">
             <div>
               <CardTitle className="text-base">
-                {loading ? "Loading…" : `${total} open ${total === 1 ? "advisory" : "advisories"}`}
+                {loading ? (
+                  <Skeleton className="h-5 w-40" />
+                ) : (
+                  `${total} open ${total === 1 ? "advisory" : "advisories"}`
+                )}
               </CardTitle>
               {total > 0 && (
                 <CardDescription className="flex flex-wrap gap-x-4 gap-y-1 pt-1 text-xs">
@@ -175,7 +180,18 @@ export function VulnerabilitiesPage() {
           </div>
         </CardHeader>
         <CardContent className="px-0 pb-0">
-          {!loading && total === 0 ? (
+          {loading ? (
+            <div className="space-y-3 px-6 py-4">
+              {Array.from({ length: 5 }).map((_, i) => (
+                <div key={i} className="flex items-center gap-4">
+                  <Skeleton className="h-5 w-16" />
+                  <Skeleton className="h-4 flex-1 max-w-[200px]" />
+                  <Skeleton className="hidden h-4 w-20 md:block" />
+                  <Skeleton className="h-4 w-32" />
+                </div>
+              ))}
+            </div>
+          ) : total === 0 ? (
             <div className="flex flex-col items-center gap-2 px-6 py-12 text-center">
               <ShieldAlert className="h-8 w-8 text-muted-foreground" />
               <p className="text-sm font-medium">No known vulnerabilities</p>
